[REFACTOR] AllergyCuration 컴포넌트 제거 후 하단에 삽입
